feat(orb_fleet): Remove unused functions o agent_service after code review comments
