+++ fleet/agent_service.go
@@ -10,13 +10,10 @@ package fleet
 
 import (
 	"context"
-	"fmt"
 	"github.com/mainflux/mainflux"
 	mfsdk "github.com/mainflux/mainflux/pkg/sdk/go"
 	"github.com/ns1labs/orb/fleet/backend"
-	"github.com/ns1labs/orb/fleet/backend/pktvisor"
 	"github.com/ns1labs/orb/pkg/errors"
-	"github.com/ns1labs/orb/pkg/types"
 	"go.uber.org/zap"
 )
 
@@ -196,137 +193,4 @@ func (svc fleetService) ViewAgentBackend(ctx context.Context, token string, name
 		return backend.GetBackend(name).Metadata(), nil
 	}
 	return nil, errors.ErrNotFound
-}
-
-func (svc fleetService) ViewAgentBackendHandler(ctx context.Context, token string, name string) (types.Metadata, error) {
-	_, err := svc.identify(token)
-	if err != nil {
-		return nil, err
-	}
-	if backend.HaveBackend(name) {
-		bk, err := pktvisor.Handlers()
-		if err != nil {
-			return nil, err
-		}
-		return bk, nil
-	}
-	return nil, errors.ErrNotFound
-}
-
-func (svc fleetService) ViewAgentBackendInput(ctx context.Context, token string, name string) (types.Metadata, error) {
-	_, err := svc.identify(token)
-	if err != nil {
-		return nil, err
-	}
-	if backend.HaveBackend(name) {
-		bk, err := pktvisor.Inputs()
-		if err != nil {
-			return nil, err
-		}
-		return bk, nil
-	}
-	return nil, errors.ErrNotFound
-}
-
-func (svc fleetService) ViewAgentBackendTaps(ctx context.Context, token string, name string) ([]BackendTaps, error) {
-	ownerID, err := svc.identify(token)
-	if err != nil {
-		return nil, err
-	}
-	if !backend.HaveBackend(name) {
-		return nil, errors.ErrNotFound
-	}
-	metadataList, err := svc.agentRepo.RetrieveAgentMetadataByOwner(ctx, ownerID)
-	if err != nil {
-		return nil, err
-	}
-
-	var list []types.Metadata
-	for _, mt := range metadataList {
-		extractTaps(mt, &list)
-	}
-
-	res, err := toBackendTaps(list)
-	if err != nil {
-		return nil, err
-	}
-	tapsGroup := groupTaps(res)
-
-	return tapsGroup, nil
-}
-
-// Used to get the taps from policy json
-func extractTaps(mt map[string]interface{}, list *[]types.Metadata) {
-	for k, v := range mt {
-		if k == "taps" {
-			m, _ := v.(map[string]interface{})
-			*list = append(*list, m)
-		} else {
-			m, _ := v.(map[string]interface{})
-			extractTaps(m, list)
-		}
-	}
-}
-
-// Used to cast the map[string]interface for a concrete struct
-func toBackendTaps(list []types.Metadata) ([]BackendTaps, error) {
-	var bkTaps []BackendTaps
-	for _, tc := range list {
-		bkTap := BackendTaps{}
-		var idx int
-		for k, v := range tc {
-			bkTap.Name = k
-			m, ok := v.(map[string]interface{})
-			if !ok {
-				return nil, errors.New("Error to group taps")
-			}
-			for k, v := range m {
-				if k == "config" {
-					m, ok := v.(map[string]interface{})
-					if !ok {
-						return nil, errors.New("Error to group taps")
-					}
-					for k, _ := range m {
-						bkTap.ConfigPredefined = append(bkTap.ConfigPredefined, []string{k}...)
-					}
-				} else {
-					bkTap.InputType = k
-				}
-			}
-			idx++
-			bkTap.TotalAgents += uint64(idx)
-			bkTaps = append(bkTaps, bkTap)
-		}
-	}
-	return bkTaps, nil
-}
-
-// Used to aggregate and sumarize the taps and return a slice of BackendTaps
-func groupTaps(taps []BackendTaps) []BackendTaps {
-	//TODO sort taps before group
-	tapsMap := make(map[string]BackendTaps)
-	for _, tap := range taps {
-		key := key(tap.Name, tap.InputType)
-		if v, ok := tapsMap[key]; ok {
-			v.ConfigPredefined = append(v.ConfigPredefined, tap.ConfigPredefined...)
-			v.TotalAgents += 1
-			tapsMap[key] = v
-		} else {
-			tapsMap[key] = BackendTaps{
-				Name:             tap.Name,
-				InputType:        tap.InputType,
-				ConfigPredefined: tap.ConfigPredefined,
-				TotalAgents:      tap.TotalAgents,
-			}
-		}
-	}
-	var bkTaps []BackendTaps
-	for _, v := range tapsMap {
-		bkTaps = append(bkTaps, v)
-	}
-	return bkTaps
-}
-
-func key(name string, inputType string) string {
-	return fmt.Sprintf("%s-%s", name, inputType)
 }
