+++ fleet/api/http/requests.go
@@ -177,3 +177,14 @@ func (req *listResourcesReq) validate() error {
 
 	return nil
 }
+
+type listAgentBackendsReq struct {
+	token string
+}
+
+func (req *listAgentBackendsReq) validate() error {
+	if req.token == "" {
+		return errors.ErrUnauthorizedAccess
+	}
+	return nil
+}
