+++ fleet/api/http/responses.go
@@ -189,3 +189,19 @@ func (s validateAgentRes) Headers() map[string]string {
 func (s validateAgentRes) Empty() bool {
 	return false
 }
+
+type agentBackendsRes struct {
+	Backends []interface{} `json:"backends,omitempty"`
+}
+
+func (s agentBackendsRes) Code() int {
+	return http.StatusOK
+}
+
+func (s agentBackendsRes) Headers() map[string]string {
+	return map[string]string{}
+}
+
+func (s agentBackendsRes) Empty() bool {
+	return false
+}
